pyproject.toml
[project]
name = "shipspy"
description = "processing of ship campaign data"
authors = [
    {name = "Laura Köhler", email = "laura.koehler@mpimet.mpg.de"},
]
dynamic = ["version"]

dependencies = [
    "xarray",
    "numpy",
    "metpy",
    "scipy",
    "pyyaml",
    "netcdf4",
    "datetime",
    "dask",
    "pint_xarray",
]


[project.optional-dependencies]
tests = [
    "pytest",
    "black"
]

[project.scripts]
shipspy = "shipspy.cli:main"

[tool.setuptools]
packages = ["shipspy"]

[tool.setuptools.dynamic]
version = {attr = "shipspy.__version__"}

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]

[build-system]
requires = ["setuptools>=45"]
build-backend = "setuptools.build_meta"
